# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.ChangeUserPasswordRequest:
    properties:
      newPassword:
        minLength: 6
        type: string
      password:
        minLength: 6
        type: string
    required:
    - newPassword
    - password
    type: object
...
  dto.CreateUserRequest:
    properties:
      email:
        type: string
      name:
        minLength: 3
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  dto.DeleteUserRequest:
    properties:
      password:
        minLength: 6
        type: string
    required:
    - password
    type: object
...
  dto.GetUserResponse:
    properties:
      email:
        type: string
      id:
        type: string
      language:
        type: string
      name:
        type: string
      theme:
        type: string
    required:
    - email
    - id
    - language
    - name
    - theme
    type: object
...
  dto.LoginUserRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  dto.Tokens:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
...
  dto.UpdateUserRequest:
    properties:
      email:
        type: string
      language:
        type: string
      name:
        type: string
      theme:
        type: string
    type: object
...
  errors.HTTPError:
    properties:
      message:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    name: Deve-Lite
  title: DashboardX API
  version: "1.0"
paths:
...
  /users/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.LoginUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Tokens'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.HTTPError'
      summary: Login a user
      tags:
      - Users
  /users/me:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Delete data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.DeleteUserRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.HTTPError'
      security:
      - BearerAuth: []
      summary: Delete a logged in user
      tags:
      - Users
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetUserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.HTTPError'
      security:
      - BearerAuth: []
      summary: Get a logged in user
      tags:
      - Users
    patch:
      consumes:
      - application/json
      parameters:
      - description: Update data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.HTTPError'
      security:
      - BearerAuth: []
      summary: Update a logged in user
      tags:
      - Users
  /users/me/password:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Change passoword data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.ChangeUserPasswordRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.HTTPError'
      security:
      - BearerAuth: []
      summary: Change a password of logged in user
      tags:
      - Users
  /users/refresh:
    post:
      consumes:
      - application/json
      description: Requires a valid refresh token sent in the Authorization header
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Tokens'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.HTTPError'
      security:
      - BearerAuth: []
      summary: Generate a new pair of user tokens
      tags:
      - Users
  /users/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Register input
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/errors.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.HTTPError'
      summary: Register a new user
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"